requirements.txt
flask
gunicorn
simple_youtube_api
python_dotenv
WooCommerce